+++ release.nix
@@ -2,8 +2,8 @@
 , nixpkgs ? (import nixpkgsBootstrap {}).fetchFromGitHub {
     owner = "NixOS";
     repo = "nixpkgs";
-    rev = "c6356ce381cd04c676eb0f6028e1f8395bdba323"; # release-17.09
-    sha256 = "16qh1zplkzswwp49j9cly7ashz1xby2qgzkchi7prn3dwjhsych1"; }
+    rev = "2231575a7ddd768fbce7d8f0c75926e5f1ee5c95"; # release-17.09
+    sha256 = "0w55wvv55j88b3hzhjsi19g8n7vl4mng701qa38g1p3gjqsqqjsv"; }
 }:
 
 let